CORE-45052: Make super category work in V3.

+++ docroot/modules/custom/alshaya_rcs/modules/alshaya_rcs_super_category/src/Service/RcsSuperCategoryHelper.php
@@ -84,17 +84,32 @@ class RcsSuperCategoryHelper {
    */
   public function syncSuperCategories() {
     foreach ($this->languageManager->getLanguages() as $language_code => $language) {
+      $this->logger->notice('Synchronizing super categories for language @language.', [
+        '@language' => $language_code,
+      ]);
+
+      $this->languageManager->setConfigOverrideLanguage($language);
+
       // Fetch categories from mdc.
       $super_categories = $this->apiWrapper->getCategories($language_code);
-      if (!empty($super_categories)) {
-        $super_categories = $this->processSuperCategories($super_categories['children_data']);
-        // Fetch existing L1 rcs categories.
-        $existing_super_categories = $this->entityTypeManager->getStorage('taxonomy_term')->loadTree('rcs_category',0, 1, TRUE);
-        // Delete orphan terms.
-        $this->deleteOrphansCategories($super_categories, $existing_super_categories);
-        // Sync categories.
-        $this->synchronizeCategories($super_categories, $existing_super_categories, $language_code);
+      if (empty($super_categories)) {
+        $this->logger->notice('No categories available in API response for language @language.', [
+          '@language' => $language_code,
+        ]);
+
+        continue;
       }
+
+      $super_categories = $this->processSuperCategories($super_categories['children_data']);
+
+      // Fetch existing L1 rcs categories.
+      $existing_super_categories = $this->entityTypeManager->getStorage('taxonomy_term')->loadTree('rcs_category',0, 1, TRUE);
+
+      // Delete orphan terms.
+      $this->deleteOrphansCategories($super_categories, $existing_super_categories);
+
+      // Sync categories.
+      $this->synchronizeCategories($super_categories, $existing_super_categories, $language_code);
     }
   }
 
@@ -108,6 +123,7 @@ class RcsSuperCategoryHelper {
    */
   protected function deleteOrphansCategories($super_categories, $existing_super_categories) {
     $default_term_id = $this->configFactory->get('rcs_placeholders.settings')->get('category.placeholder_tid');
+
     foreach ($existing_super_categories as $category) {
       // Ignore placeholder taxonomy term.
       if ($category->id() === $default_term_id) {
@@ -115,11 +131,17 @@ class RcsSuperCategoryHelper {
       }
       // Delete term if not found in mdc categories and is not level 2.
       $slug = $category->get('field_category_slug')->getString();
-      if (count(explode('/', $slug)) < 2
+      if ($slug
+        && count(explode('/', trim($slug, '/'))) < 2
         && !in_array($slug, array_keys($super_categories))
-        && !in_array($slug . '/', array_keys($super_categories))
-      ) {
+        && !in_array($slug . '/', array_keys($super_categories))) {
+        // @todo delete the advanced page too if available.
         $category->delete();
+
+        $this->logger->warning('Category with id @id and slug @slug deleted as no longer available in API response.', [
+          '@id' => $category->id(),
+          '@slug' => $slug,
+        ]);
       }
     }
   }
@@ -136,8 +158,10 @@ class RcsSuperCategoryHelper {
    */
   protected function synchronizeCategories($super_categories, $existing_super_categories, $lang_code) {
     $term_storage = $this->entityTypeManager->getStorage('taxonomy_term');
-    // Map existing cateogeries with their slug.
+
+    // Map existing categories with their slug.
     $existing_super_categories = $this->processExistingCategories($existing_super_categories, $lang_code);
+
     foreach ($super_categories as $slug => $category) {
       // Check if rcs category with the slug already exists.
       $existing_category = NULL;
@@ -148,60 +172,63 @@ class RcsSuperCategoryHelper {
         $existing_category = $existing_super_categories[$slug . '/'];
       }
 
-      if (!empty($existing_category)) {
-        if ($existing_category->hasTranslation($lang_code)) {
-          continue;
-        }
-        $term = $existing_category->addTranslation($lang_code, [
-          'vid' => 'rcs_category',
-          'name' => $category['name'],
-          'langcode' => $lang_code,
-        ]);
-      }
-      else {
+      if (empty($existing_category)) {
         $term = $term_storage->create([
           'vid' => 'rcs_category',
           'name' => $category['name'],
           'langcode' => $lang_code,
         ]);
       }
-      $term->get('field_category_slug')->setValue($category['url_key']);
+      else {
+        if (!$existing_category->hasTranslation($lang_code)) {
+          $term = $existing_category->addTranslation($lang_code, [
+            'vid' => 'rcs_category',
+            'name' => $category['name'],
+            'langcode' => $lang_code,
+          ]);
+        }
+        else {
+          $term = $existing_category->getTranslation($lang_code);
+        }
+      }
+
+      $term->get('field_category_slug')->setValue($category['url_path']);
       $term->get('field_commerce_id')->setValue($category['id']);
+      $term->get('weight')->setValue($category['position']);
       $term->save();
 
-      // Create Advanced pages.
-      $url_keys = [$category['url_key'], $category['url_key'] . '/'];
+      // Create / Update Advanced pages.
+      $url_keys = [$category['url_path'], $category['url_path'] . '/'];
       $node_storage = $this->entityTypeManager->getStorage('node');
       $query = $node_storage->getQuery();
       $query->condition('type', 'advanced_page');
       $query->condition('field_category_slug', $url_keys, 'IN');
       $nodes = $query->execute();
       if (empty($nodes)) {
-        $node = $node_storage
-          ->create(
-            [
-              'title' => $category['name'],
-              'type' => 'advanced_page',
-              'langcode' => $lang_code,
-            ]
-          );
+        $node = $node_storage->create([
+          'title' => $category['name'],
+          'type' => 'advanced_page',
+          'langcode' => $lang_code,
+        ]);
       }
       else {
         $node = $node_storage->load(current($nodes));
-        if ($node instanceof NodeInterface && $node->hasTranslation($lang_code)) {
-          continue;
+        if ($node instanceof NodeInterface) {
+          if ($node->hasTranslation($lang_code)) {
+            $node = $node->getTranslation($lang_code);
+          }
+          else {
+            $node = $node->addTranslation($lang_code, [
+              'title' => $category['name'],
+              'type' => 'advanced_page',
+              'langcode' => $lang_code,
+            ]);
+          }
         }
-        $node = $node->addTranslation($lang_code,
-          [
-            'title' => $category['name'],
-            'type' => 'advanced_page',
-            'langcode' => $lang_code,
-          ]
-        );
       }
 
       $node->get('field_use_as_department_page')->setValue(TRUE);
-      $node->get('field_category_slug')->setValue($category['url_key']);
+      $node->get('field_category_slug')->setValue($category['url_path']);
       $node->save();
     }
   }
@@ -223,18 +250,24 @@ class RcsSuperCategoryHelper {
   }
 
   /**
-   * Process mdc super category attributes.
+   * Process MDC categories to add required attributes.
+   *
+   * @param array $categories
+   *   Categories to process.
+   *
+   * @return array
+   *   Processed category data.
    */
-  protected function processSuperCategories($super_categories) {
+  protected function processSuperCategories(array $categories): array {
     $processed_categories = [];
-    foreach ($super_categories as $category) {
-      $url_key = $this->getCustomAttribute($category, 'url_key');
-      $processed_categories[$url_key] = [
-        'name' => $category['name'],
-        'url_key' => $url_key,
-        'id' => $category['id'],
-      ];
+    foreach ($categories as $category) {
+      $category['url_path'] = $this->getCustomAttribute($category, 'url_path');
+      $category['field_logo_active_image'] = $this->getCustomAttribute($category, 'field_logo_active_image');
+      $category['field_logo_inactive_image'] = $this->getCustomAttribute($category, 'field_logo_inactive_image');
+      $category['field_logo_header_image'] = $this->getCustomAttribute($category, 'field_logo_header_image');
+      $processed_categories[$category['url_path']] = $category;
     }
+
     return $processed_categories;
   }
